refactor(work): extract kind-checks-shared-base to eliminate jscpd clones
Both work-code-checker and work-completion-checker kind-checks/shared.js
carried byte-identical readFile/readChangedFiles/spec re-export blocks;
extracted to plugins/work/scripts/workflows/lib/kind-checks-shared-base.js.
Both consumers now import from the shared base, eliminating duplicate-block
debt. Behavior unchanged; completion-checker + code-checker tests pass.
