+++ plugins/work/scripts/workflows/work-completion-checker/lib/kind-checks/shared.js
@@ -10,52 +10,14 @@
 
 const fs = require('node:fs');
 const path = require('node:path');
-const { spawnSync } = require('node:child_process');
 
 const specShared = require('../../../work-spec/lib/kind-checks/shared');
-const config = require('../../../lib/config');
+const {
+  readFile,
+  readChangedFiles,
+  specSharedReexports,
+} = require('../../../lib/kind-checks-shared-base');
 const { extractRequirementIdsFromBulletBlock } = require('../../../lib/requirement-ids');
-
-function readFile(p) {
-  try {
-    return fs.readFileSync(p, 'utf8');
-  } catch {
-    return null;
-  }
-}
-
-/**
- * Get the changed-file list. Prefer the snapshot pr-context.json (written
- * by pr-next.js diff_audit) so completion-checker sees the same diff the
- * PR phase locked in. Fall back to git diff if absent.
- */
-function readChangedFiles(ctx) {
-  const ctxPath = path.join(ctx.tasksDir, 'pr-context.json');
-  if (fs.existsSync(ctxPath)) {
-    try {
-      const j = JSON.parse(fs.readFileSync(ctxPath, 'utf8'));
-      if (Array.isArray(j.files)) return j.files.slice();
-    } catch {
-      /* fall through */
-    }
-  }
-  const root = ctx.worktreeRoot || process.cwd();
-  // Honor BASE_BRANCH / symbolic-ref so dev-based repos don't fall back to
-  // origin/main (which is behind merges and surfaces phantom files).
-  for (const base of config.getDiffBaseCandidates({ cwd: root })) {
-    const r = spawnSync('git', ['diff', '--name-only', `${base}...HEAD`], {
-      cwd: root,
-      encoding: 'utf8',
-    });
-    if (r.status === 0) {
-      return r.stdout
-        .split('\n')
-        .map((s) => s.trim())
-        .filter(Boolean);
-    }
-  }
-  return [];
-}
 
 /**
  * Walk `## Task N — <title>` blocks and synthesize coverage rows from each
@@ -302,19 +264,5 @@ module.exports = {
   readSuggestedScopeFiles,
   readTestReport,
   // Re-exports from spec-side shared:
-  readBrief: specShared.readBrief,
-  readSpec: specShared.readSpec,
-  readTasks: specShared.readTasks,
-  sliceSection: specShared.sliceSection,
-  filesInFilesToModify: specShared.filesInFilesToModify,
-  detectKinds: specShared.detectKinds,
-  MalformedTasksError: specShared.MalformedTasksError,
-  preflightTasksManifest: specShared.preflightTasksManifest,
-  briefForbidsBackend: specShared.briefForbidsBackend,
-  isBackendFile: specShared.isBackendFile,
-  isFrontendFile: specShared.isFrontendFile,
-  isE2eFile: specShared.isE2eFile,
-  isDevopsFile: specShared.isDevopsFile,
-  isAppSourceFile: specShared.isAppSourceFile,
-  KIND_NAMES: specShared.KIND_NAMES,
+  ...specSharedReexports(specShared),
 };
